rename: 폴더명 변경(/notice -> agree)
